soften the spotlight and bug fix
Fixed bug where the ambient light was getting added even for a spotlight of 0 intensity.

Softened the spotlight a bit. The amount is hardcoded atm (with the 100) so it's up to you if you want to make that customisable at some point. The lower the multiplier the more softness you get.